requirement.txt
absl-py==0.15.0
alabaster==0.7.12
astunparse==1.6.3
Babel==2.10.3
build==0.8.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.1.0
commonmark==0.9.1
cycler==0.11.0
Cython==0.29.30
docopt==0.6.2
docutils==0.17.1
filelock==3.8.0
flatbuffers==1.12
fonttools==4.35.0
gast==0.3.3
gensim==3.8.3
glove-python-binary==0.2.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
huggingface-hub==0.8.1
idna==3.3
imagesize==1.4.1
importlib-metadata==4.12.0
Jinja2==3.1.2
joblib==1.1.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
llvmlite==0.39.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
numba==0.56.0
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pep517==0.13.0
Pillow==9.2.0
pip==22.1.2
pipreqs==0.4.11
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
recommonmark==0.7.1
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
sacremoses==0.0.53
scikit-learn==0.24.2
scipy==1.7.3
seaborn==0.11.2
setuptools==61.2.0
shap==0.41.0
six==1.15.0
skfeature==1.0.0
slicer==0.0.7
smart-open==6.0.0
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinx-markdown-tables==0.0.17
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.10.3
tomli==2.0.1
torch==1.8.1+cu102
torchaudio==0.8.1
torchvision==0.9.1+cu102
tqdm==4.64.0
transformers==4.12.5
typing-extensions==3.7.4.3
urllib3==1.26.11
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.12.1
yarg==0.1.9
yellowbrick==1.3
zipp==3.8.1
